% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market.R
\name{cost_values}
\alias{cost_values}
\alias{total_cost}
\alias{variable_cost}
\alias{marginal_cost}
\alias{average_cost}
\alias{average_variable_cost}
\alias{min_average_cost}
\title{Cost at a level of output}
\usage{
total_cost(cost, q, ...)

variable_cost(cost, q, ...)

marginal_cost(cost, q, ...)

average_cost(cost, q, ...)

average_variable_cost(cost, q, ...)

min_average_cost(cost, q_max = NULL, ...)
}
\arguments{
\item{cost}{A cost object from \code{\link[=quadratic_cost]{quadratic_cost()}} or \code{\link[=production_cost]{production_cost()}}.}

\item{q}{Output. May be a vector.}

\item{...}{Passed on to methods.}

\item{q_max}{Upper bound for the numerical search in
\code{min_average_cost()}; only needed for a \code{\link[=production_cost]{production_cost()}}.}
}
\value{
A numeric vector, except \code{min_average_cost()}, which returns a
list with the output \code{q} at which average cost is lowest and the average
cost \code{ac} there. For a \code{\link[=quadratic_cost]{quadratic_cost()}} with no fixed cost or no
quadratic term average cost has no interior minimum; \code{q} is then \code{NA}
and \code{ac} the limiting value.
}
\description{
Cost at a level of output
}
\examples{
cst <- quadratic_cost(fixed = 100, a = 20, b = 0.5)
total_cost(cst, c(0, 10, 20))
marginal_cost(cst, 10)     # 20 + 2 * 0.5 * 10
average_cost(cst, 10)
min_average_cost(cst)       # q = sqrt(F / b)
}
