% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masthead.R
\name{print.econ_plot}
\alias{print.econ_plot}
\alias{plot.econ_plot}
\title{Draw an econ_plot}
\usage{
\method{print}{econ_plot}(x, newpage = TRUE, ...)

\method{plot}{econ_plot}(x, ...)
}
\arguments{
\item{x}{An \code{econ_plot} from \code{\link[=econ_masthead]{econ_masthead()}}.}

\item{newpage}{Start a new grid page first?}

\item{...}{Ignored, for compatibility with \code{\link[=print]{print()}}.}
}
\value{
\code{x}, invisibly.
}
\description{
Draw an econ_plot
}
