% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{scale_econ}
\alias{scale_econ}
\alias{scale_colour_econ}
\alias{scale_color_econ}
\alias{scale_fill_econ}
\alias{scale_colour_econ_c}
\alias{scale_color_econ_c}
\alias{scale_fill_econ_c}
\title{Economist-style colour and fill scales}
\usage{
scale_colour_econ(palette = "main", reverse = FALSE, ...)

scale_color_econ(palette = "main", reverse = FALSE, ...)

scale_fill_econ(palette = "main", reverse = FALSE, ...)

scale_colour_econ_c(palette = "blues", reverse = FALSE, ...)

scale_color_econ_c(palette = "blues", reverse = FALSE, ...)

scale_fill_econ_c(palette = "blues", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Palette name, see \code{\link[=econ_pal]{econ_pal()}}.}

\item{reverse}{Reverse the colour order?}

\item{...}{Passed to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}} or
\code{\link[ggplot2:scale_gradient]{ggplot2::scale_colour_gradientn()}}.}
}
\value{
A ggplot2 scale.
}
\description{
Discrete scales draw from the categorical palette in order; continuous
scales interpolate a sequential or diverging ramp.
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(wt, mpg, colour = factor(cyl))) +
  geom_point(size = 3) +
  scale_colour_econ() +
  theme_econ()
}
